requirements.txt
flask
flask-sqlalchemy
tweepy 
basilica 
python-decouple 
python-dotenv 
scikit-learn 
gunicorn 
psycopg2
numpy
google-cloud-bigquery
textblob
pandas